Update question_views.py
add if state to check user's uploading state and prevent multiple uploading.

+++ ept/views/question_views.py
@@ -46,46 +46,61 @@ def create():
     form = QuestionForm()
 
     if form.validate_on_submit():
-        subject = form.subject.data
-        file_content = form.fileContent.data
-        
-        uploadfilepath = UPLOAD_FOLDER+"/"+str(uuid.uuid4())
-        file_content.save(uploadfilepath)
-        tika = Tika(uploadfilepath)
+        if current_user.uploading_state:  # Check whether user is uploading
+            flash("You are already uploading a file. Please wait for the current upload to finish.")
+            return redirect(url_for('main.index'))
 
-        file_type = tika.get_type()
-        print(file_type)
-        
-        # file type check
-        if not allowed_file(file_type):
-            flash("Only PDF or HTML or TXT or PNG or MSOffice file formats are allowed.")
-            if os.path.isfile(uploadfilepath): os.remove(uploadfilepath)
-            return render_template('question/question_form.html', form=form)
-
-        # file text token check
-        file_content_text = tika.get()
-        token_count = Token(file_content_text).gpt4_token_count()
-        if os.path.isfile(uploadfilepath): os.remove(uploadfilepath)
-        if token_count > TOKEN_LIMIT:
-            flash("The document contains "+str(token_count)+" tokens. ExamGPT limits text in documents to "+str(TOKEN_LIMIT)+" tokens.")
-            return render_template('question/question_form.html', form=form)
-        
-        # file text content check
-        contentcheck = ContentCheck(file_content_text)
-        contentcheck_string = contentcheck.get()
-        contentcheck_data = json.loads(contentcheck_string)
-        if(contentcheck_data['clarity'] == 'false' or contentcheck_data['problem_development'] == 'false'
-           or contentcheck_data['XSS_safe'] == 'false' or contentcheck_data['prompt_safe'] == 'false'):
-            flash(contentcheck_data['comment'])
-            return render_template('question/question_form.html', form=form)
-
-        # add db data
-        new_file = Question(subject=subject, file_text=file_content_text, create_date=datetime.now(), user=g.user)
-        db.session.add(new_file)
+        # Set the uploading state to true when the user starts uploading
+        current_user.uploading_state = True
         db.session.commit()
-        return redirect(url_for('main.index'))
-        
+
+        try:
+            subject = form.subject.data
+            file_content = form.fileContent.data
+            
+            uploadfilepath = UPLOAD_FOLDER+"/"+str(uuid.uuid4())
+            file_content.save(uploadfilepath)
+            tika = Tika(uploadfilepath)
+    
+            file_type = tika.get_type()
+            print(file_type)
+            
+            # file type check
+            if not allowed_file(file_type):
+                flash("Only PDF or HTML or TXT or PNG or MSOffice file formats are allowed.")
+                if os.path.isfile(uploadfilepath): os.remove(uploadfilepath)
+                return render_template('question/question_form.html', form=form)
+    
+            # file text token check
+            file_content_text = tika.get()
+            token_count = Token(file_content_text).gpt4_token_count()
+            if os.path.isfile(uploadfilepath): os.remove(uploadfilepath)
+            if token_count > TOKEN_LIMIT:
+                flash("The document contains "+str(token_count)+" tokens. ExamGPT limits text in documents to "+str(TOKEN_LIMIT)+" tokens.")
+                return render_template('question/question_form.html', form=form)
+            
+            # file text content check
+            contentcheck = ContentCheck(file_content_text)
+            contentcheck_string = contentcheck.get()
+            contentcheck_data = json.loads(contentcheck_string)
+            if(contentcheck_data['clarity'] == 'false' or contentcheck_data['problem_development'] == 'false'
+               or contentcheck_data['XSS_safe'] == 'false' or contentcheck_data['prompt_safe'] == 'false'):
+                flash(contentcheck_data['comment'])
+                return render_template('question/question_form.html', form=form)
+    
+            # add db data
+            new_file = Question(subject=subject, file_text=file_content_text, create_date=datetime.now(), user=g.user)
+            db.session.add(new_file)
+            db.session.commit()
+            return redirect(url_for('main.index'))
+            
+        except Exception as e:
+            flash("An error occurred while processing your request. Please try again.")
+            current_user.uploading_state = False  # Reset the uploading state in case of an error
+            db.session.commit()
+            
     return render_template('question/question_form.html', form=form)
+        
         
         
 # @bp.route('/edit/', methods=('GET', 'POST'))
